fix(finance): rehydrate finance-wire-types to match consumer expectations
Resolves the finance regression flagged across several sessions. Consumers
on main (zomes/payments, zomes/bridge, tests/bridge_test.rs, tests/currency_mint_test/*)
had been updated to use a richer wire-types shape, but the shape itself
was never pushed from the session-substrate-wiring worktree. As the user
confirmed, nobody else is actively iterating here — so the safe move was
to read consumer call-sites + struct-construction shapes from main itself
and synthesize the canonical wire-types to match.

Added 9 missing types (shapes inferred from consumer construction sites):
  ApplyDemurrageInput      { member_did, currency_id, local/regional/global_commons_pool_id }
  DemurrageResult          { deducted: u64, redistributed: bool }
  MintSapFromGovernanceInput { amount, proposal_id, recipient_did }
  BalanceResponse          { member_did, currency, balance, available }  (generic, currency-agnostic)
  DepositCollateralInput   { depositor_did, collateral_type, collateral_amount, oracle_rate }
  FeeTierResponse          { member_did, mycel_score: f64, tier_name, base_fee_rate }
  FinanceBridgeHealth      { healthy, agent, zomes }
  ProcessPaymentInput      { source_happ, from_did, to_did, amount, currency, reference }
  UpdateCollateralHealthInput { collateral_id, obligation_amount }

Reshaped 3 existing structs to match consumer construction:
  RegisterCollateralInput  — swapped to cross-hApp-addressable shape
                             (source_happ + asset_id + value_estimate + currency)
  SapBalanceResponse       — raw + effective + pending_demurrage split
                             (last_demurrage_at: i64 to match Timestamp::as_micros)
  FinanceRuntimeDiscovery  — added default_dao_did + dao_dids (Vec) for
                             multi-DAO membership; dropped legacy single-DAO
                             field no caller was setting

Also fixed 4 pre-existing `u32 vs usize` drift sites in zomes/bridge +
zomes/payments where `Option<u32> limit.unwrap_or(100)` was being passed
directly to `.take()` / `.saturating_sub()`. Surgical `as usize` casts;
doesn't change the wire contract.

Net effect: `cargo check --workspace` in mycelix-finance goes from
50+ errors (E0432 + E0560 + E0609 + E0308) to zero. Unblocks every
demo flow that touches TEND/SAP/MYCEL/treasury/staking/recognition —
all of Part 2 of MYCELIX_DEMO_READINESS.md's Kagiso storyboard that
exercises economic state now has a green build to run against.

Includes fixes for items 1 from the follow-up list ("Finance wire-types
coordination"); separate from item 2 (unified hApp install, which still
needs `nix develop`).

+++ mycelix-finance/zomes/bridge/coordinator/src/lib.rs
@@ -260,7 +260,7 @@ pub struct BroadcastFinanceEventInput {
 /// Get payment history for a DID (paginated, default limit 100)
 #[hdk_extern]
 pub fn get_payment_history(input: GetPaymentHistoryInput) -> ExternResult<Vec<Record>> {
-    let limit = input.limit.unwrap_or(100);
+    let limit = input.limit.unwrap_or(100) as usize;
     let query = LinkQuery::try_new(anchor_hash(&input.did)?, LinkTypes::DidToPayments)?;
     let links = get_links(query, GetStrategy::default())?;
